+++ library/src/main/java/com/getstream/sdk/chat/rest/Message.java
@@ -129,8 +129,6 @@ public class Message implements UserEntity {
     @TypeConverters(ExtraDataConverter.class)
     private HashMap<String, Object> extraData;
 
-    private boolean isThreadParent = false; /*Used to tag Thread Parent Message*/
-
     private boolean isStartDay = false;
     private boolean isYesterday = false;
     private boolean isToday = false;
@@ -302,7 +300,6 @@ public class Message implements UserEntity {
         clone.parentId = parentId;
         clone.command = command;
         clone.commandInfo = commandInfo;
-        clone.isThreadParent = isThreadParent;
         clone.status = status;
         return clone;
     }
@@ -498,14 +495,6 @@ public class Message implements UserEntity {
 
     public void setCommandInfo(Map<String, String> commandInfo) {
         this.commandInfo = commandInfo;
-    }
-
-    public boolean isThreadParent() {
-        return isThreadParent;
-    }
-
-    public void setThreadParent(boolean threadParent) {
-        isThreadParent = threadParent;
     }
 
     @Override